chore：时空限制改为整数存储

+++ src/main/java/io/github/yush1x/tenjudge/server/problem/dto/ProblemConfig.java
@@ -7,8 +7,8 @@ import java.util.List;
 @Data
 public class ProblemConfig {
     private String name;
-    private Double time_limit;
-    private Double memory_limit;
+    private Integer time_limit;
+    private Integer memory_limit;
     private String judge_type;
     private Integer difficulty;
     private List<String> tags;
